+++ packages/assets-controllers/src/TokenListController.test.ts
@@ -507,10 +507,7 @@ const getRestrictedMessenger = (
   const messenger = controllerMessenger.getRestricted({
     name,
     allowedActions: ['NetworkController:getNetworkClientById'],
-    allowedEvents: [
-      'TokenListController:stateChange',
-      'NetworkController:stateChange',
-    ],
+    allowedEvents: ['NetworkController:stateChange'],
   });
 
   return messenger;