+++ packages/permission-controller/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Uncategorized
+- [`base-controller`] Enforce that `RestrictedControllerMessenger` is not initialized with internal actions/events in allow lists ([#2051](https://github.com/MetaMask/core/pull/2051))
+- Deprecate `BaseControllerV1` and use `BaseControllerV2` as default ([#2078](https://github.com/MetaMask/core/pull/2078))
+- Add `wallet_revokePermissions` RPC method ([#1889](https://github.com/MetaMask/core/pull/1889))
+- Refactor controllers to use `"getState"` action, `"stateChange"` event types defined in `base-controller` ([#2029](https://github.com/MetaMask/core/pull/2029))
 
 ## [5.0.1]
 ### Changed